+++ demisto_sdk/commands/common/hook_validations/layout.py
@@ -5,6 +5,18 @@ from demisto_sdk.commands.common.tools import print_error
 
 class LayoutValidator(BaseValidator):
 
+    def is_valid_layout(self, validate_rn=True):  # type: () -> bool
+        """Check whether the layout is valid or not.
+
+        Returns:
+            bool. Whether the layout is valid or not
+        """
+        answers = [
+            super().is_valid_file(validate_rn),
+            self.is_valid_version()
+        ]
+        return all(answers)
+
     def is_valid_version(self):
         # type: () -> bool
         """Return if version is valid.
@@ -16,14 +28,3 @@ class LayoutValidator(BaseValidator):
             print_error(Errors.wrong_version(self.file_path, self.DEFAULT_VERSION))
             return False
         return True
-
-    def is_valid_layout(self):  # type: () -> bool
-        """Check whether the layout is valid or not.
-
-        Returns:
-            bool. Whether the layout is valid or not
-        """
-        answers = [
-            self.is_valid_version()
-        ]
-        return all(answers)