+++ demisto_sdk/commands/unify/unifier.py
@@ -18,12 +18,12 @@ from demisto_sdk.commands.common.constants import TYPE_TO_EXTENSION, INTEGRATION
 
 class Unifier:
 
-    def __init__(self, indir: str, dir_name=INTEGRATIONS_DIR, outdir='',
+    def __init__(self, input: str, dir_name=INTEGRATIONS_DIR, output: str = '',
                  image_prefix=DEFAULT_IMAGE_PREFIX, force: bool = False):
 
         directory_name = ""
         for optional_dir_name in DIR_TO_PREFIX:
-            if optional_dir_name in indir:
+            if optional_dir_name in input:
                 directory_name = optional_dir_name
 
         if not directory_name:
@@ -31,12 +31,12 @@ class Unifier:
                         'should contain either Integrations or Scripts directories')
 
         self.image_prefix = image_prefix
-        self.package_path = indir
+        self.package_path = input
         self.use_force = force
         if self.package_path.endswith(os.sep):
             self.package_path = self.package_path.rstrip(os.sep)
 
-        self.dest_path = outdir
+        self.dest_path = output
 
         yml_paths, self.yml_path = get_yml_paths_in_dir(self.package_path, Errors.no_yml_file(self.package_path))
         for path in yml_paths:
@@ -195,18 +195,21 @@ class Unifier:
 
     def get_code_file(self, script_type):
         """Return the first code file in the specified directory path
-        :param script_type: script type: .py or .js
+        :param script_type: script type: .py, .js, .ps1
         :type script_type: str
         :return: path to found code file
         :rtype: str
         """
 
         ignore_regex = (r'CommonServerPython\.py|CommonServerUserPython\.py|demistomock\.py|_test\.py'
-                        r'|conftest\.py|__init__\.py|ApiModule\.py|vulture_whitelist\.py')
-
+                        r'|conftest\.py|__init__\.py|ApiModule\.py|vulture_whitelist\.py'
+                        r'|CommonServerPowerShell\.ps1|CommonServerUserPowerShell\.ps1|demistomock\.ps1|\.Tests\.ps1')
+        if self.package_path.endswith('/'):
+            self.package_path = self.package_path[:-1]  # remove the last / as we use os.path.join
         if self.package_path.endswith('Scripts/CommonServerPython'):
             return os.path.join(self.package_path, 'CommonServerPython.py')
-
+        if self.package_path.endswith('Scripts/CommonServerPowerShell'):
+            return os.path.join(self.package_path, 'CommonServerPowerShell.ps1')
         if self.package_path.endswith('ApiModule'):
             return os.path.join(self.package_path, os.path.basename(os.path.normpath(self.package_path)) + '.py')
 
@@ -226,7 +229,10 @@ class Unifier:
         if module_import:
             script_code = self.insert_module_code(script_code, module_import, module_name)
 
-        clean_code = self.clean_python_code(script_code)
+        if script_type == '.py':
+            clean_code = self.clean_python_code(script_code)
+        if script_type == '.ps1':
+            clean_code = self.clean_pwsh_code(script_code)
 
         if self.is_script_package:
             if yml_data.get('script', '') not in ('', '-'):
@@ -318,4 +324,10 @@ class Unifier:
         # print function is imported in python loop
         if remove_print_future:  # docs generation requires to leave this
             script_code = script_code.replace("from __future__ import print_function", "")
+        return script_code
+
+    @staticmethod
+    def clean_pwsh_code(script_code):
+        script_code = script_code.replace(". $PSScriptRoot\\demistomock.ps1", "")
+        script_code = script_code.replace(". $PSScriptRoot\\CommonServerPowerShell.ps1", "")
         return script_code